+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "python-library"
-version = "2.2.8"
+version = "2.3.0"
 description = "python library"
 requires-python = ">=3.11"
 dependencies = [
