+++ src/python_library/thread/queue_thread.py
@@ -4,7 +4,34 @@ from threading import Lock
 
 from python_library.job.job import IJob
 from python_library.job_queue.job_queue import IJobQueue, JobQueue
-from python_library.thread.thread import abThread, IQueueThread
+from python_library.thread.thread import IThread, abThread
+
+
+class IQueueThread(IThread):
+
+    @abstractmethod
+    def set_shared_job_queue(self, shared_job_queue: IJobQueue, shared_job_queue_lock: Lock) -> None: ...
+
+    @abstractmethod
+    def push_shared_job_queue(self, job: IJob) -> None: ...
+
+    @abstractmethod
+    def pop_shared_job_queue(self) -> IJob | None: ...
+
+    @abstractmethod
+    def size_shared_job_queue(self) -> int: ...
+
+    @abstractmethod
+    def set_shared_queue(self, shared_queue: Dict[str, IJobQueue], shared_queue_lock: Dict[str, Lock]) -> None: ...
+
+    @abstractmethod
+    def push_shared_queue(self, name: str, job: IJob) -> None: ...
+
+    @abstractmethod
+    def pop_shared_queue(self, name: str) -> Optional[IJob]: ...
+
+    @abstractmethod
+    def size_shared_queue(self, name: str) -> int: ...
 
 
 class QueueThread(abThread, IQueueThread):
